backend/pyproject.toml
[project]
name = "academic-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "feedparser>=6.0.12",
    "openai>=2.30.0",
    "pdfplumber>=0.11.9",
    "pymupdf>=1.27.2.2",
    "requests>=2.33.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4",
    "httpx[socks]>=0.28.1",
]
